chore: move csv_row into its own model
