+++ backend/tests/test_beneficiary.py
@@ -1,10 +1,10 @@
 from datetime import date
 
-from api.db.crud.beneficiary import get_beneficiary_from_csv
+from api.db.crud.beneficiary import get_beneficiary_from_personal_information
 from api.db.crud.wanted_job import find_wanted_job_for_beneficiary
 from api.db.models.beneficiary import Beneficiary
 from cdb_csv import pe
-from cdb_csv.csv_row import PrincipalCsvRow
+from cdb_csv.models.csv_row import PrincipalCsvRow
 
 
 async def test_get_beneficiary_with_wanted_jobs(pe_principal_csv_series, db_connection):
@@ -13,8 +13,11 @@ async def test_get_beneficiary_with_wanted_jobs(pe_principal_csv_series, db_conn
     _, series = next(pe_principal_csv_series.iterrows())
     csv_row: PrincipalCsvRow = await pe.map_principal_row(series)
 
-    beneficiary: Beneficiary | None = await get_beneficiary_from_csv(
-        db_connection, csv_row
+    beneficiary: Beneficiary | None = await get_beneficiary_from_personal_information(
+        db_connection,
+        firstname=csv_row.prenom,
+        lastname=csv_row.nom,
+        birth_date=csv_row.date_naissance,
     )
 
     assert beneficiary is not None
@@ -43,8 +46,13 @@ async def test_get_beneficiary_without_wanted_jobs(
 
             csv_row: PrincipalCsvRow = await pe.map_principal_row(series)
 
-            beneficiary: Beneficiary | None = await get_beneficiary_from_csv(
-                db_connection, csv_row
+            beneficiary: Beneficiary | None = (
+                await get_beneficiary_from_personal_information(
+                    db_connection,
+                    firstname=csv_row.prenom,
+                    lastname=csv_row.nom,
+                    birth_date=csv_row.date_naissance,
+                )
             )
 
             assert beneficiary is not None
@@ -71,8 +79,13 @@ async def test_get_beneficiary_without_notebook(pe_principal_csv_series, db_conn
 
             csv_row: PrincipalCsvRow = await pe.map_principal_row(series)
 
-            beneficiary: Beneficiary | None = await get_beneficiary_from_csv(
-                db_connection, csv_row
+            beneficiary: Beneficiary | None = (
+                await get_beneficiary_from_personal_information(
+                    db_connection,
+                    firstname=csv_row.prenom,
+                    lastname=csv_row.nom,
+                    birth_date=csv_row.date_naissance,
+                )
             )
 
             assert beneficiary is not None
@@ -91,8 +104,13 @@ async def test_get_beneficiary_with_unknown_rome_code(
 
             csv_row: PrincipalCsvRow = await pe.map_principal_row(series)
 
-            beneficiary: Beneficiary | None = await get_beneficiary_from_csv(
-                db_connection, csv_row
+            beneficiary: Beneficiary | None = (
+                await get_beneficiary_from_personal_information(
+                    db_connection,
+                    firstname=csv_row.prenom,
+                    lastname=csv_row.nom,
+                    birth_date=csv_row.date_naissance,
+                )
             )
 
             assert beneficiary is not None
